+++ ladybug/src/main/java/nl/nn/ibistesttool/DeploymentSpecificsBeanPostProcessor.java
@@ -47,7 +47,8 @@ public class DeploymentSpecificsBeanPostProcessor implements BeanPostProcessor {
 				}
 				appConstants.setProperty("testtool.enabled", testToolEnabled);
 			}
-			//testTool.setReportGeneratorEnabled(testToolEnabled); // no longer necessary to switch off report generator, when debugger advice is disabled
+			// enable/disable testtool via two switches, until one of the switches has become deprecated
+			testTool.setReportGeneratorEnabled(testToolEnabled); 
 			IbisDebuggerAdvice.setEnabled(testToolEnabled);
 		}
 		if (bean instanceof nl.nn.testtool.storage.file.Storage) {