+++ ladybug/src/main/java/nl/nn/ibistesttool/LadybugPipe.java
@@ -110,10 +110,12 @@ public class LadybugPipe extends FixedForwardPipe {
 		reportRunner.setSecurityContext(new IbisSecurityContext(session, checkRoles));
 		boolean reportGeneratorEnabledOldValue = testTool.getReportGeneratorEnabled();
 		if(enableReportGenerator) {
+			IbisDebuggerAdvice.setEnabled(true);
 			testTool.setReportGeneratorEnabled(true);
 		}
 		reportRunner.run(reports, false, true);
 		if(enableReportGenerator) {
+			IbisDebuggerAdvice.setEnabled(reportGeneratorEnabledOldValue);
 			testTool.setReportGeneratorEnabled(reportGeneratorEnabledOldValue);
 		}